requirements.txt
flask 
requests 
openpyxl==3.1.5 
beautifulsoup4==4.12.3 
pillow==10.3.0 
playwright 
pymssql==2.3.2 
python-dotenv
npm install [email protected]
npx playwright install
Flask-CORS


# sudo apt update
# sudo apt install -y libnss3 libatk1.0-0 libatk-bridge2.0-0 libcups2 libxkbcommon0 libgbm-dev libpangocairo-1.0-0
# python -m playwright install-deps
# sudo mount -o remount,size=2G /dev/shm
# sudo apt install xvfb
# xvfb-run -a python app.py
# chmod -R 755 ~/igwestudy/ig_web_robot/myenv/bin

# pip uninstall playwright
# pip install playwright
# python -m playwright install
